:recycle: Split configs into separate files
